refactor: Update all dependencies

+++ shared/src/commonMain/kotlin/app/octocon/app/utils/time.kt
@@ -1,6 +1,5 @@
 package app.octocon.app.utils
 
-import kotlinx.datetime.Clock
 import kotlinx.datetime.DateTimeUnit
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.Month
@@ -8,11 +7,12 @@ import kotlinx.datetime.TimeZone
 import kotlinx.datetime.plus
 import kotlinx.datetime.toLocalDateTime
 import kotlinx.datetime.until
+import kotlin.time.Clock
 
 fun monthDays(year: Int, month: Month): Int {
   val start = LocalDate(year, month, 1)
   val end = start.plus(1, DateTimeUnit.MONTH)
-  return start.until(end, DateTimeUnit.DAY)
+  return start.until(end, DateTimeUnit.DAY).toInt()
 }
 
 typealias MonthYearPair = Pair<Int, Int>
